+++ src/player-util.h
@@ -124,5 +124,6 @@ void player_handle_post_move(struct player *p, bool eval_trap,
 void disturb(struct player *p);
 void search(struct player *p);
 bool player_has_monster_in_view(const struct player *p);
+void player_forget_the_world(struct player *p);
 
 #endif /* !PLAYER_UTIL_H */